Hack a fix for traceback on "unmerged" files in a merge/rebase
My understanding of this is mostly absent, but I *think* this restores the old
behaviour.  I figure a working tool is more useful than one that doesn't run at
all!

Apparently triggered by a change in the behaviour of git checkout-index.  I
guess they changed it to be more explicit about what you want it to do in these
complicated situations.