Don't try to build manpage on master
I think I intended to have the manpage on master too, probably so people could
install from there (as opposed to from a release package or release branch) and
have the manpage.

Might be awkward to achieve with this branching structure I used in this project
though because there will be merge conflicts on the generated manpage.  I could
work around that somehow or kill the complicated branching, but for now I'm just
removing it instead since seems it never got built on master.