README.md: Update installation instructions
These days, installing globally via `pip` generally doesn't work. I think `uv` is a very good option for people these days, but `pipx` is also good.

It took me a while to discover that this is surprisingly easy, in spite of the fact that `git-meld-index` doesn't seem to be on PyPi. Hopefully, adding it to the README would save people some time.